Changed all doubles to ints in files

+++ Character_header/Character.h
@@ -7,26 +7,26 @@
 using namespace std;
 
 enum CharType{MAINCHAR, GOBLIN, DRAGON, NPC};
-const double MAX_HEALTH = 200.0;
+const int MAX_HEALTH = 200;
 
 class Character : public Object {
 protected:
-double health;
-double attackAmount;
-double damage_amount;
+int health;
+int attackAmount;
+int damage_amount;
 string name;
 CharType type;
 
 public:
 Character(){};
-Character(CharType, const string &, double, double){};
+Character(CharType, const string &, int, int){};
 
-double getHealth () const {return health;};
+int getHealth () const {return health;};
 CharType getType() {return type;};
-virtual void setHealth(double)= 0;
+virtual void setHealth(int)= 0;
 virtual bool isalive() const =0;
-virtual void damage(double dam) = 0;
-virtual void takeDamage(double) = 0 ;
+virtual void damage(int dam) = 0;
+virtual void takeDamage(int) = 0 ;
 virtual void attack(Character &) = 0;
 
 };
